chore: lib: apply compilation fixes

+++ libs/pkcs11/src/CSP/VerificaConCIE.h
@@ -9,13 +9,12 @@
 #define LIBS_PKCS11_SRC_CSP_VERIFICACONCIE_H_
 
 #include <stdio.h>
-#include <sys/types.h>
 
-#include "CSP/AbilitaCIE.h"
+#include "AbilitaCIE.h"
 #include "Sign/CIEVerify.h"
 
 typedef CK_RV (*verificaConCIEfn)(const char* inFilePath);
 typedef CK_RV (*getNumberOfSignfn)(void);
-typedef CK_RV (*getVerifyInfofn)(u_int64_t index, struct verifyInfo_t* vInfos);
+typedef CK_RV (*getVerifyInfofn)(int index, struct verifyInfo_t* vInfos);
 typedef CK_RV (*estraiP7mfn)(const char* inFilePath, const char* outFilePath);
 #endif  // LIBS_PKCS11_SRC_CSP_VERIFICACONCIE_H_
